fix(core): move the Spring beans under modules/ into the scan roots
Three beans lived under modules/, in packages outside the two roots the platform
treats as its scan convention (org.eclipse.dirigible.components and
.engine - the pair DataSourceSystemConfig uses for @EnableJpaRepositories and
dirigible.scan.packages). build/application never noticed, because
DirigibleApplication sits in org.eclipse.dirigible and its default scan covers
the whole tree; any assembly that NAMES its scan packages loses them.

All three are contributor-shaped, so their absence is silent. A
WebSocketConfigurer that is never registered just means the endpoint 404s; a
List<ConnectionEnhancer> or List<DatabaseConfigurator> injection point just
receives a shorter list. No missing-bean failure, no warning - the context comes
up clean and the capability is quietly gone.

  HanaConnectionEnhancer   -> components/data/data-sources
  HanaDatabaseConfigurator -> components/data/data-sources
  ConsoleWebsocketConfig   -> components/ide/ide-logs

The HANA pair moves next to the injection points that consume it
(DirigibleDataSourceFactory, DataSourceInitializer) and joins its SPI siblings
under components/; both are package-private, so nothing outside their package
could reference them and the move costs no pom change - data-sources already
depends on core-database, api-security and Hikari, and already carries a
vendor-specific dependency (database-sql-h2).

ConsoleWebsocketConfig joins the four other WebSocketConfigurers, every one of
which already lives under components/ (ide-terminal, ide-java-lsp,
ide-java-debug, data-transfer) - it was the only one stranded in modules/.
ide-logs takes a dependency on commons-resources for ConsoleWebsocketHandler.
The appender, handler and record deliberately stay where they are: logback
configuration names ConsoleLoggingAppender by fully qualified class name, in
this repo and in every downstream one, so moving that package would break
configurations we do not control.

Also drops the vestigial @Component from DirigibleSourceProvider. Every consumer
constructs it directly, so the stereotype only advertised an injection point
that never existed - and its removal is behaviour-neutral.

No Spring bean stereotype remains anywhere under modules/, which restores the
invariant the tree is described by ("pure libraries with no Spring wiring"). The
one remaining Spring reference there is ConsoleWebsocketHandler, which extends
TextWebSocketHandler but is not a bean - it is constructed by the config and
called statically by the appender.

Registering the beans through an @AutoConfiguration was the other option in the
report, and was rejected: an auto-configuration class under
org.eclipse.dirigible is itself covered by DirigibleApplication's default scan,
so in the full assembly it would be registered twice - once by the scan and once
from the imports file - giving two WebSocketConfigurers and a duplicated handler
registration. Moving has no such ambiguity.

Closes #6635
