+++ components/data/data-sources/src/main/java/org/eclipse/dirigible/components/data/sources/manager/HanaConnectionEnhancer.java
@@ -7,7 +7,7 @@
  *
  * SPDX-FileCopyrightText: Eclipse Dirigible contributors SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.database.sql.dialects.hana;
+package org.eclipse.dirigible.components.data.sources.manager;
 
 import org.eclipse.dirigible.components.api.security.UserFacade;
 import org.eclipse.dirigible.components.database.ConnectionEnhancer;
